# docs/swagger.yaml
basePath: /
host: localhost:4123
definitions:
  model.Application:
    properties:
      c_name:
        type: string
      created_at:
        type: string
      deployments:
        items:
          $ref: '#/definitions/model.ApplicationDeployment'
        type: array
      description:
        type: string
      e_name:
        type: string
      health_check_type:
        type: string
      health_check_url:
        type: string
      id:
        type: integer
      listen_port:
        type: integer
      repo:
        allOf:
        - $ref: '#/definitions/model.Repo'
        description: 修正外键关系定义
      repo_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.ApplicationDeployment:
    properties:
      application:
        allOf:
        - $ref: '#/definitions/model.Application'
        description: 关联关系
      application_id:
        type: integer
      build_source:
        type: string
      build_template:
        $ref: '#/definitions/model.BuildTemplate'
      build_template_id:
        type: integer
      created_at:
        type: string
      deployment_template:
        $ref: '#/definitions/model.DeploymentTemplate'
      deployment_template_id:
        type: integer
      description:
        type: string
      environment:
        $ref: '#/definitions/model.Environment'
      environment_id:
        type: integer
      id:
        type: integer
      platform_credential_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.ApplicationDeploymentRequest:
    properties:
      application_id:
        type: integer
      build_source:
        type: string
      build_template_id:
        type: integer
      deployment_template_id:
        type: integer
      description:
        type: string
      environment_id:
        type: integer
      platform_credential_id:
        type: integer
    type: object
  model.ApplicationDeploymentResponse:
    properties:
      application_id:
        type: integer
      build_source:
        type: string
      build_template_id:
        type: integer
      created_at:
        type: string
      deployment_template_id:
        type: integer
      description:
        type: string
      environment_id:
        type: integer
      id:
        type: integer
      platform_credential_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.ApplicationRequest:
    properties:
      c_name:
        type: string
      description:
        type: string
      e_name:
        type: string
      health_check_type:
        type: string
      health_check_url:
        type: string
      listen_port:
        type: integer
      repo_id:
        type: integer
    type: object
  model.ApplicationResponse:
    properties:
      c_name:
        type: string
      created_at:
        type: string
      deployment_count:
        description: 部署配置数量
        type: integer
      description:
        type: string
      e_name:
        type: string
      health_check_type:
        type: string
      health_check_url:
        type: string
      id:
        type: integer
      listen_port:
        type: integer
      repo_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.BuildTemplate:
    properties:
      created_at:
        type: string
      creator:
        type: string
      dockerfile:
        type: string
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      pipeline:
        type: string
      repos:
        description: 关联Repo表的多对多关系
        items:
          $ref: '#/definitions/model.Repo'
        type: array
      updated_at:
        type: string
      updater:
        type: string
    type: object
...
  model.DeploymentTemplate:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator:
        type: string
      description:
        type: string
      display_name:
        type: string
      histories:
        description: 正确的关联历史记录关系
        items:
          $ref: '#/definitions/model.DeploymentTemplateHistory'
        type: array
      id:
        type: integer
      name:
        type: string
      parameters:
        type: string
      repos:
        description: 关联仓库表的多对多关系
        items:
          $ref: '#/definitions/model.Repo'
        type: array
      status:
        type: string
      template_type:
        type: string
      updated_at:
        type: string
      updater:
        type: string
      variables:
        type: string
      version:
        type: string
    type: object
  model.DeploymentTemplateHistory:
    properties:
      change_reason:
        type: string
      content:
        type: string
      created_at:
        type: string
      deployment_template_id:
        type: integer
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      modifier:
        type: string
      name:
        type: string
      parameters:
        type: string
      template_type:
        type: string
      variables:
        type: string
      version:
        type: string
    type: object
...
  model.Environment:
    properties:
      config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Repo:
    properties:
      applications:
        items:
          $ref: '#/definitions/model.Application'
        type: array
      c_name:
        type: string
      created_at:
        type: string
      deployment_templates:
        description: 添加与部署模板的多对多关联关系
        items:
          $ref: '#/definitions/model.DeploymentTemplate'
        type: array
      e_name:
        type: string
      id:
        type: integer
      platform:
        type: string
      repo_build_path:
        type: string
      repo_department:
        type: string
      repo_deploy_type:
        type: string
      repo_desc:
        type: string
      repo_language:
        type: string
      repo_manager:
        type: string
      repo_number:
        type: string
      repo_ssh_url:
        type: string
      repo_url:
        type: string
      templates:
        description: 添加与构建模板的多对多关联关系
        items:
          $ref: '#/definitions/model.BuildTemplate'
        type: array
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Minimal OpenAPI spec for Zebra-CICD endpoints
  title: Zebra-CICD API
  version: 0.1.0
paths:
  /api/application/template:
    get:
      description: 根据应用服务ID获取部署配置列表
      parameters:
      - description: 应用服务ID
        in: query
        name: application_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationDeployment'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取应用部署配置列表
      tags:
      - application-deployments
    post:
      consumes:
      - application/json
      description: 为应用服务创建部署配置
      parameters:
      - description: 部署配置信息
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/model.ApplicationDeploymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ApplicationDeploymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建应用部署配置
      tags:
      - application-deployments
  /api/application/template/{id}:
    delete:
      description: 根据ID删除部署配置
      parameters:
      - description: 部署配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除应用部署配置
      tags:
      - application-deployments
    get:
      description: 根据部署配置ID获取详细信息
      parameters:
      - description: 部署配置ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ApplicationDeploymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取应用部署配置
      tags:
      - application-deployments
    put:
      consumes:
      - application/json
      description: 根据ID更新部署配置信息
      parameters:
      - description: 部署配置ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部署配置信息
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/model.ApplicationDeploymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ApplicationDeploymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新应用部署配置
      tags:
      - application-deployments
  /api/application/template/environment:
    get:
      description: 根据环境ID获取该环境下所有的部署配置
      parameters:
      - description: 环境ID
        in: query
        name: environment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationDeployment'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据环境获取部署配置列表
      tags:
      - application-deployments
  /api/applications:
    get:
      description: 根据仓库ID获取应用服务列表
      parameters:
      - description: 仓库ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取应用服务列表
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: 基于仓库创建应用服务
      parameters:
      - description: 应用服务信息
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/model.ApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建应用服务
      tags:
      - applications
  /api/applications/{id}:
    delete:
      description: 根据ID删除应用服务及其相关部署配置
      parameters:
      - description: 应用服务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除应用服务
      tags:
      - applications
    get:
      description: 根据应用服务ID获取详细信息
      parameters:
      - description: 应用服务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取应用服务
      tags:
      - applications
    put:
      consumes:
      - application/json
      description: 根据ID更新应用服务信息
      parameters:
      - description: 应用服务ID
        in: path
        name: id
        required: true
        type: integer
      - description: 应用服务信息
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/model.ApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新应用服务
      tags:
      - applications
...
swagger: "2.0"